Relocate Interactive Librarian to Common
